LPS-44048 Use integer for counters

+++ portal-impl/src/com/liferay/portal/util/PropsValues.java
@@ -1842,7 +1842,7 @@ public class PropsValues {
 
 	public static final int TRASH_ENTRIES_MAX_AGE = GetterUtil.getInteger(PropsUtil.get(PropsKeys.TRASH_ENTRIES_MAX_AGE));
 
-	public static final long TRASH_ENTRIES_MAX_CLEAN_COUNT = GetterUtil.getLong(PropsUtil.get(PropsKeys.TRASH_ENTRIES_MAX_CLEAN_COUNT), -1L);
+	public static final int TRASH_ENTRIES_MAX_CLEAN_COUNT = GetterUtil.getInteger(PropsUtil.get(PropsKeys.TRASH_ENTRIES_MAX_CLEAN_COUNT), -1);
 
 	public static final int TRASH_ENTRY_CHECK_INTERVAL = GetterUtil.getInteger(PropsUtil.get(PropsKeys.TRASH_ENTRY_CHECK_INTERVAL));
